+++ src/blocks/builders/mod.rs
@@ -7,6 +7,7 @@ mod file;
 mod header;
 mod image;
 mod input;
+mod rich_text;
 mod section;
 mod video;
 
@@ -17,5 +18,6 @@ pub use file::FileBuilder;
 pub use header::HeaderBuilder;
 pub use image::ImageBuilder;
 pub use input::InputBuilder;
+pub use rich_text::RichTextBuilder;
 pub use section::SectionBuilder;
 pub use video::VideoBuilder;
