+++ src/composition_objects/builders/mod.rs
@@ -17,5 +17,7 @@ pub use dispatch_action_configuration::DispatchActionConfigurationBuilder;
 pub use input_parameter::InputParameterBuilder;
 pub use option::OptBuilder;
 pub use option_group::OptGroupBuilder;
+pub use slack_file::SlackFileBuilder;
 pub use text::TextBuilder;
 pub use trigger::TriggerBuilder;
+pub use workflow::WorkflowBuilder;
