+++ src/rich_text_elements/types/link.rs
@@ -1,15 +1,15 @@
 use super::CodableStyle;
 use serde::Serialize;
 
-/// [**text**](https://api.slack.com/reference/block-kit/blocks#text-element-type)
+/// [**link**](https://api.slack.com/reference/block-kit/blocks#link-element-type)
 /// type of [Rich text element types](https://api.slack.com/reference/block-kit/blocks#element-types)
 ///
 /// # Example
 ///
 /// ```
-/// # use slack_messaging::blocks::rich_text::elements::types::{RichTextElementTypeText, CodableStyle};
-/// let text = RichTextElementTypeText::builder()
-///     .text("hello")
+/// # use slack_messaging::rich_text_elements::types::{RichTextElementTypeLink, CodableStyle};
+/// let link = RichTextElementTypeLink::builder()
+///     .url("https://google.com")
 ///     .style(
 ///         CodableStyle::builder()
 ///             .bold(true)
@@ -18,23 +18,29 @@ use serde::Serialize;
 ///     .build();
 ///
 /// let expected = serde_json::json!({
-///     "type": "text",
-///     "text": "hello",
+///     "type": "link",
+///     "url": "https://google.com",
 ///     "style": {
 ///         "bold": true
 ///     }
 /// });
 ///
-/// let json = serde_json::to_value(text).unwrap();
+/// let json = serde_json::to_value(link).unwrap();
 ///
 /// assert_eq!(json, expected);
 /// ```
 #[derive(Debug, Clone, Serialize)]
-pub struct RichTextElementTypeText {
+pub struct RichTextElementTypeLink {
     #[serde(rename = "type")]
     pub(super) kind: &'static str,
 
-    pub(super) text: String,
+    pub(super) url: String,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub(super) text: Option<String>,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub(super) r#unsafe: Option<bool>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub(super) style: Option<CodableStyle>,
