+++ src/rich_text_elements/types/text.rs
@@ -1,15 +1,15 @@
 use super::CodableStyle;
 use serde::Serialize;
 
-/// [**link**](https://api.slack.com/reference/block-kit/blocks#link-element-type)
+/// [**text**](https://api.slack.com/reference/block-kit/blocks#text-element-type)
 /// type of [Rich text element types](https://api.slack.com/reference/block-kit/blocks#element-types)
 ///
 /// # Example
 ///
 /// ```
-/// # use slack_messaging::blocks::rich_text::elements::types::{RichTextElementTypeLink, CodableStyle};
-/// let link = RichTextElementTypeLink::builder()
-///     .url("https://google.com")
+/// # use slack_messaging::rich_text_elements::types::{RichTextElementTypeText, CodableStyle};
+/// let text = RichTextElementTypeText::builder()
+///     .text("hello")
 ///     .style(
 ///         CodableStyle::builder()
 ///             .bold(true)
@@ -18,29 +18,23 @@ use serde::Serialize;
 ///     .build();
 ///
 /// let expected = serde_json::json!({
-///     "type": "link",
-///     "url": "https://google.com",
+///     "type": "text",
+///     "text": "hello",
 ///     "style": {
 ///         "bold": true
 ///     }
 /// });
 ///
-/// let json = serde_json::to_value(link).unwrap();
+/// let json = serde_json::to_value(text).unwrap();
 ///
 /// assert_eq!(json, expected);
 /// ```
 #[derive(Debug, Clone, Serialize)]
-pub struct RichTextElementTypeLink {
+pub struct RichTextElementTypeText {
     #[serde(rename = "type")]
     pub(super) kind: &'static str,
 
-    pub(super) url: String,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(super) text: Option<String>,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(super) r#unsafe: Option<bool>,
+    pub(super) text: String,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub(super) style: Option<CodableStyle>,
